bsp: multiconfig: k3r5: allow customization by donwstream layers
This patch adds the possibility to change some bitbake variable
that cannot be changed anywhere else, one of then is the TMPDIR.

Using the same TMPDIR it is very sensitive and prone to several errors
when used in more complex situations.
This configuration forces that all native packages have to be the same between all
machines and this requirement is very easy to break.
Suppose you use a macinhe override somewhere on a native recipe
and this requirement is no longer met.
Many of these anomalies can be verified with the use of the rm_work and
create-spdx bitbake classes.

A previous attempt [1] had already been made but was refused
but this way it can be done without side effects for other users.

[1] https://lists.yoctoproject.org/g/meta-ti/message/14767

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>