conf: switch k3r5 multiconfig to baremetal
There are couple of k3r5 components that need to be built for the target
and they are self-contained and don't use glibc. Moreover, when building
a cross-compile toolchain for k3r5, there's no need to build glibc and
all its dependencies. While build-time speed up is nice, the main benefit
comes when packaging a nativesdk toolchain for distribution, reducing the
number of packages and the resulting size of the toolchain installer.

To avoid potential namespace conflict in package feeds between nativesdk
binary packages of Linux glibc and k3r5 baremetal toolchains, adjust the
package suffix accordingly.

This also requires a rather impactful change in the deployment structure,
which affects CIs and user experience. By default different TCLIBC builds
(e.g. glibc and baremetal in this case) are built in separate TMPDIRs to
avoid conflicts, especially on the nativesdk side. Also the default config
has DEPLOY_DIR residing under corresponding TMPDIR. Of course, this can be
changed by local.conf or <distro>.conf, which meta-ti-bsp has no control
over. Unlike multiconfig-image-packager example that can access deployed
artifacts across multiconfigs to package in an image, we use wic and its
IMAGE_BOOT_FILES list to package a final image. Unfortunately, wic cannot
access deployed artifacts across multiconfigs in separate TMPDIRs, as it
uses DEPLOY_DIR_IMAGE to collect the artifacts. To overcome this, we need
to set up a "common" deploy directory for images and sdks produced by
multiconfigs, which is located outside of TMPDIRs, set by TI_COMMON_DEPLOY
variable that can be changed or adjusted by distros as needed.

Signed-off-by: Denys Dmytriyenko <denys@konsulko.com>